asdsad

+++ CollegeERPSystem/src/main/java/com/erp/repositories/FacultyRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.erp.model.Course;
 import com.erp.model.Faculty;
 
-public interface FacultyRepository extends JpaRepository<Faculty, Integer>{
+public interface FacultyRepository extends JpaRepository<Faculty, Long>{
 	
 	Faculty findAllById(Long id);
